+++ docs/topology-status.md
@@ -14,7 +14,7 @@ Generated by `devtools render-topology-status`. Reads `docs/plans/topology-targe
 | archive-action-event | — | 4 | 4 | 0 | 0 |
 | archive-artifact-taxonomy | — | 4 | 4 | 0 | 0 |
 | archive-attachment | — | 2 | 2 | 0 | 0 |
-| archive-conversation | — | 10 | 10 | 0 | 0 |
+| archive-conversation | — | 11 | 11 | 0 | 0 |
 | archive-filter | archive filter semantics | 4 | 4 | 0 | 0 |
 | archive-message | — | 7 | 7 | 0 | 0 |
 | archive-phase | — | 2 | 2 | 0 | 0 |
@@ -31,7 +31,7 @@ Generated by `devtools render-topology-status`. Reads `docs/plans/topology-targe
 
 - **Stable** (no move scoped): 582
 - **Kernel** (polylogue/ root): 10
-- **Primitives** (lib-root + storage-root): 13
+- **Primitives** (lib-root + storage-root): 12
 - **TBD** (cell needs explicit assignment): 0
 - **Total declared**: 695
 - **Realized polylogue/**/*.py**: 695 files declared
