+++ polylogue/products/archive_rollups.py
@@ -7,8 +7,8 @@ from collections.abc import Iterable, Sequence
 from dataclasses import dataclass, field
 from datetime import UTC, datetime
 
+from polylogue.archive.conversation.repo_identity import normalize_repo_names
 from polylogue.archive.session.session_profile import SessionProfile
-from polylogue.lib.repo_identity import normalize_repo_names
 from polylogue.products.archive import (
     SessionTagRollupProduct,
     profile_bucket_day,
